--- Assets/Other/Controls.inputactions
@@ -1,33 +1,0 @@
-{
-    "name": "Controls",
-    "maps": [
-        {
-            "name": "UI",
-            "id": "391797bd-4897-4b76-9846-c3f9332c6673",
-            "actions": [
-                {
-                    "name": "Pause",
-                    "type": "Button",
-                    "id": "5506e558-5707-43e6-aec8-762690ea086a",
-                    "expectedControlType": "Button",
-                    "processors": "",
-                    "interactions": ""
-                }
-            ],
-            "bindings": [
-                {
-                    "name": "",
-                    "id": "e8a6a839-b6f7-4f38-a322-10e005377fd4",
-                    "path": "<Keyboard>/escape",
-                    "interactions": "",
-                    "processors": "",
-                    "groups": "",
-                    "action": "Pause",
-                    "isComposite": false,
-                    "isPartOfComposite": false
-                }
-            ]
-        }
-    ],
-    "controlSchemes": []
-}
