+++ Assets/Scripts/BaseCode.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2fe319bae0051f5498b183766b85248c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
