+++ Assets/Scripts/BaseCode/Entity.cs
@@ -19,14 +19,3 @@ public class Entity : MonoBehaviour
         GameObject.Destroy(this.gameObject, 0);
     }
 }
-
-[CreateAssetMenu(fileName = "EntityData", menuName = "LTG5/Entities/EntityData", order = 0)]
-public class EntityData : ScriptableObject{
-
-    public enum elementTypes{None, Earth, Fire, Water, Air};
-    public elementTypes element;
-    
-    public float MaxHealth;
-    public float manaDropAmount;
-
-}
