+++ Assets/Scripts/BaseCode/Entity.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eee29da1c7428a548b3d1e48e844528b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
