+++ Assets/Scripts/BaseCode/EntityData.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 058afafa72ec9b84b86ae4654ac2134c, type: 3}
+  m_Name: EntityData
+  m_EditorClassIdentifier: 
+  element: 0
+  MaxHealth: 0
+  manaDropAmount: 0
