+++ Assets/Scripts/BaseCode/EntityData.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bfba6d0e118ecff458d1d1f23518d59f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
