+++ Assets/Scripts/BaseCode/EntityData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 058afafa72ec9b84b86ae4654ac2134c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
