+++ Assets/Scripts/BaseCode/PlatformerController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 20036a77c63e0b847a885ba07bf97bac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
