+++ Assets/Scripts/BaseCode/PlatformerParameters.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b1cea7653e06a2c42bbfd030f2ab6431
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
