+++ Assets/Scripts/BaseCode/PlatformerParameters.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8714dea5dd8605f4dac94d57b029f0cd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
